backend/requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-jwt-extended==4.4.4
flask-cors==3.0.10
python-dotenv==1.0.0
email-validator==1.3.1
pillow==9.4.0
boto3==1.26.84
pytest==7.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5